requirements.txt
# ep-py-digital-twin-framework dependencies
# Generated from production venv, pruned to direct dependencies.
# Install with: pip install -r requirements.txt
#
# NOTE: pyenergyplus is bundled with EnergyPlus and is not pip-installable.
# Add your EnergyPlus installation directory to PYTHONPATH before running.
#
# NOTE: seeq and seeq-spy are commercial packages tied to a licensed Seeq server.
# Version numbers below reflect the production deployment -- newer versions
# may require testing before use.

# OPC-UA
asyncua==1.1.0

# PostgreSQL
# Use psycopg2 (source build) on Linux/Mac or psycopg2-binary on Windows
psycopg2-binary==2.9.11

# Data handling
numpy==1.24.3
pandas==2.0.3

# Solar position calculations (used in custom/conversion.py)
pvlib==0.15.0

# HTTP requests (support scripts and retrieval)
requests==2.32.5

# Machine learning (used in OPC predictor and future ML integration)
scikit-learn==1.3.0
scipy==1.11.1
h5py==3.16.0

# Database ORM (SQLAlchemy used by persistence layer)
SQLAlchemy==2.0.48

# Cryptography (asyncua dependency)
cryptography==46.0.5

# Date/time utilities
python-dateutil==2.9.0.post0
pytz==2026.1.post1

# Scheduling (used in support_scripts ingestion services)
# Note: not in production venv freeze -- install separately if using support scripts
schedule>=1.2.0

# Seeq historian client (commercial -- requires licensed Seeq server)
# See: https://python-seeq.readthedocs.io/
seeq==66.10.0.20250206
seeq-spy==193.20
